Added message to profile page in case user gives a wrong sk_username

+++ app/models/user.rb
@@ -82,9 +82,13 @@ class User < ActiveRecord::Base
   def save_calendar_entries_from_api
     uri = "http://api.songkick.com/api/3.0/users/#{sk_username}/calendar.json?reason=attendance&apikey=hackday"
     response = HTTParty.get(uri)
-    CalendarEntry.where(user_id: self.id).destroy_all
-    response['resultsPage']['results']['calendarEntry'].each do |entry|
-      CalendarEntry.find_or_create_by(CalendarEntry.parse_calendar_entry(entry, self.id))
+    if (response['resultsPage']['status'] == "error")
+      return false
+    elsif
+      CalendarEntry.where(user_id: self.id).destroy_all
+      response['resultsPage']['results']['calendarEntry'].each do |entry|
+        CalendarEntry.find_or_create_by(CalendarEntry.parse_calendar_entry(entry, self.id))
+      end
     end
   end
